+++ site_settings/constants.py
@@ -1,10 +1,10 @@
 # The website version number
-WEBSITE_VERSION = "1.2.8"
-WEBSITE_VERSION_DATE = "2021-05-06"
+WEBSITE_VERSION = "2.0.0-beta"
+WEBSITE_VERSION_DATE = "2021-10-05"
 
 WEBSITE_FULL_VERSION = "{} - {}".format(
 	WEBSITE_VERSION, WEBSITE_VERSION_DATE
 )
 
 # Update this to force reload of cached css
-CSS_VERSION = "1.1"
+CSS_VERSION = "1.0"